fix the parsing error of top rated places(3.4 Int to Double) fix the pagination issues in the top rated places

+++ Arounduaetest/Arounduaetest/Helper/Model/CitiesPlaces/CitiesPlaces.swift
@@ -27,7 +27,7 @@ struct PlacesData : Decodable {
 struct Places : Decodable {
     let _id : String?
     let title : Title?
-    let averageRating : Int?
+    let averageRating : Double?
     let images : [Images]?
 }
